requirements.txt
aiohttp>=3.9
python-dotenv>=1.0
pydantic>=2.7          # type-safe models
gspread>=6.0
google-auth>=2.29
PyYAML>=6.0
tenacity>=9.1          # retries w/ backoff
streamlit>=1.32.0      # for the web UI
pandas>=2.0.0          # for data manipulation
google-generativeai>=0.3.0  # Gemini API
